# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OneStorage",
    "version": "0.3.0",
    "description": "Container-free Git hosting on Cloudflare Workers. Independent API; not wire-compatible with third-party SDKs."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/repos/{namespace}/{repo}/base": {
      "delete": {
        "operationId": "unset_base_repo",
        "summary": "Unset Base Repo",
        "description": "See the OneStorage API contract in the source archive (docs/API.md).",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "URL-encoded repository name; grouped names occupy one path segment."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "repo:write",
        "responses": {
          "200": {
            "description": "Operation succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Validation, permission, conflict or capacity error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/repos/{namespace}/{repo}/git-credentials": {
      "post": {
        "operationId": "create_generic_git_credential",
        "summary": "Create Git Credential",
        "description": "See the OneStorage API contract in the source archive (docs/API.md).",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "URL-encoded repository name; grouped names occupy one path segment."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "repo:write",
        "responses": {
          "201": {
            "description": "Operation succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Validation, permission, conflict or capacity error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 10000
                  }
                },
                "required": [
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      },
      "put": {
        "operationId": "update_generic_git_credential",
        "summary": "Update Git Credential",
        "description": "See the OneStorage API contract in the source archive (docs/API.md).",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "URL-encoded repository name; grouped names occupy one path segment."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "repo:write",
        "responses": {
          "200": {
            "description": "Operation succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Validation, permission, conflict or capacity error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 10000
                  }
                },
                "required": [
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/repos/{namespace}/{repo}/git-credentials/{id}": {
      "delete": {
        "operationId": "delete_generic_git_credential",
        "summary": "Delete Git Credential",
        "description": "See the OneStorage API contract in the source archive (docs/API.md).",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "URL-encoded repository name; grouped names occupy one path segment."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "repo:write",
        "responses": {
          "200": {
            "description": "Operation succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Validation, permission, conflict or capacity error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/repos/{namespace}/{repo}/pull-upstream": {
      "post": {
        "operationId": "pull_upstream",
        "summary": "Pull Upstream",
        "description": "See the OneStorage API contract in the source archive (docs/API.md).",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "URL-encoded repository name; grouped names occupy one path segment."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-scope": "git:write",
        "responses": {
          "202": {
            "description": "Operation succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Validation, permission, conflict or capacity error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "PAT or client-signed JWT; scopes are independent."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      }
    }
  }
}